Email strategy setted. Log in route added. Code refactoring.
