+++ back/src/models/index.js
@@ -1,0 +1,3 @@
+module.exports = function(sequelize) {
+    require('./User.js')(sequelize);
+}
