--- back/src/models/piJulianToDelete/Diet.js
@@ -1,11 +1,0 @@
-const { DataTypes } = require('sequelize');
-
-module.exports = function(sequelize) {
-  sequelize.define('Diet', {
-    name: {
-      type: DataTypes.STRING,
-      unique: true,
-      allowNull: false, 
-    },
-  });
-};
