+++ back/src/passport/index.js
@@ -1,6 +1,6 @@
 const passport = require('passport');
-const {User} = require('./models/User.js').models;
-require('./mailStrategy.js');
+const {User} = require('../db.js').models;
+require('./emailStrategy.js');
 
 passport.serializeUser(function(user, done) {
   done(null, user.id);
@@ -8,7 +8,7 @@ passport.serializeUser(function(user, done) {
 
 passport.deserializeUser(async function(id, done) {
   try {
-    const user = await User.findById(id)
+    const user = await User.findByPk(id)
     done(null, user);
   } catch(err) {
       return done(err);
