+++ back/src/routes/index.js
@@ -1,18 +1,4 @@
-const passport = require('')
-const { Router } = require('express');
-const { User } = require("../db");
-// Importar todos los routers;
-// Ejemplo: const authRouter = require('./auth.js');
-const postUser = require("./users/postUser");
-const putUser = require("./users/putUser");
-
-const router = Router();
-
-// Configurar los routers
-// Ejemplo: router.use('/auth', authRouter);
-router.post("/createUser", postUser);
-router.put("/users/:id", putUser);
-
-
-
-module.exports = router;
+module.exports = function(app) {
+    app.use('/users', require('./users'))
+    app.use('/session', require('./session'))
+};
