+++ back/src/routes/session/index.js
@@ -1,0 +1,6 @@
+const router = require('express').Router();
+const localLogin = require('./localLogin.js');
+
+router.post('/login', localLogin);
+
+module.exports = router;
